SearchSpecifications are JPA-specific, and thus cannot move to Framework. We settle for objects-jpa/.../util
